requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2024.10
data-wizard==2.0.0
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-heroku==0.3.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
gunicorn==23.0.0
html-json-forms==1.1.1
idna==3.10
itertable==2.2.0
natural-keys==2.1.1
openpyxl==3.1.5
packaging==24.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==65.5.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
whitenoise==6.9.0
